Match the right $_ under v6.d
A regex in boolean context per v6.d captures the $_ at the point where
the regex literal appears. These two lines will stop working in
upcoming rakudo releases. There are two ways to fix it. One is to put
`use v6.c` in appropriate places, which will provide previous
behavior. Another is to match $_ explicitly so that it works for all
versions. The latter approach is used here.